+++ cookbooks/dnsupdate-nativex/attributes/default.rb
@@ -1,12 +1,12 @@
 #
 # Internal Domain Name
-default['nativex-dnsupdate']['int_domain'] = 'nativexintern.com'
+default['dnsupdate-nativex']['int_domain'] = 'nativexintern.com'
 
 # internal ZoneID
-default['nativex-dnsupdate']['int_zone_id'] = 'Z1NZHUXT02UQJ'
+default['dnsupdate-nativex']['int_zone_id'] = 'Z1NZHUXT02UQJ'
 
 # External Domain Name
-default['nativex-dnsupdate']['ext_domain'] = 'nativexintern.com'
+default['dnsupdate-nativex']['ext_domain'] = 'nativexintern.com'
 
 # external ZoneID
-default['nativex-dnsupdate']['ext_zone_id'] = 'Z1NZHUXT02UQJ'
+default['dnsupdate-nativex']['ext_zone_id'] = 'Z1NZHUXT02UQJ'